Split miners.github.precomputed_prs module and wrap loaders funcs into classes
