+++ server/athenian/api/controllers/miners/github/precomputed_prs/open_prs.py
@@ -1,0 +1,200 @@
+from datetime import datetime, timezone
+from typing import Collection, Dict, Iterable, Mapping, Tuple
+
+import databases
+import numpy as np
+import pandas as pd
+import sentry_sdk
+from sqlalchemy import and_, insert, select
+from sqlalchemy.dialects.postgresql import insert as postgres_insert
+
+from athenian.api.controllers.miners.github.precomputed_prs.utils import \
+    append_activity_days_filter, collect_activity_days
+from athenian.api.controllers.miners.types import MinedPullRequest, PullRequestFacts
+from athenian.api.models.metadata.github import PullRequest
+from athenian.api.models.precomputed.models import GitHubOpenPullRequestFacts
+from athenian.api.tracing import sentry_span
+
+
+class OpenPRFactsLoader:
+    """Loader for open PRs facts."""
+
+    @classmethod
+    @sentry_span
+    async def load_open_pull_request_facts(cls,
+                                           prs: pd.DataFrame,
+                                           account: int,
+                                           pdb: databases.Database,
+                                           ) -> Dict[str, PullRequestFacts]:
+        """
+        Fetch precomputed facts about the open PRs from the DataFrame.
+
+        We filter open PRs inplace so the user does not have to worry about that.
+        """
+        open_indexes = np.nonzero(prs[PullRequest.closed_at.key].isnull().values)[0]
+        node_ids = prs.index.take(open_indexes)
+        authors = dict(zip(node_ids, prs[PullRequest.user_login.key].values[open_indexes]))
+        ghoprf = GitHubOpenPullRequestFacts
+        default_version = ghoprf.__table__.columns[ghoprf.format_version.key].default.arg
+        selected = [
+            ghoprf.pr_node_id,
+            ghoprf.pr_updated_at,
+            ghoprf.repository_full_name,
+            ghoprf.data,
+        ]
+        rows = await pdb.fetch_all(
+            select(selected)
+            .where(and_(ghoprf.pr_node_id.in_(node_ids),
+                        ghoprf.format_version == default_version,
+                        ghoprf.acc_id == account)))
+        if not rows:
+            return {}
+        found_node_ids = [r[0].rstrip() for r in rows]
+        found_updated_ats = [r[1] for r in rows]
+        if pdb.url.dialect == "sqlite":
+            found_updated_ats = [dt.replace(tzinfo=timezone.utc) for dt in found_updated_ats]
+        updated_ats = prs[PullRequest.updated_at.key].take(open_indexes)
+        passed_mask = updated_ats[found_node_ids] <= found_updated_ats
+        passed_node_ids = set(updated_ats.index.take(np.where(passed_mask)[0]))
+        facts = {}
+        for row in rows:
+            node_id = row[ghoprf.pr_node_id.key]
+            if node_id in passed_node_ids:
+                facts[node_id] = PullRequestFacts(
+                    data=row[ghoprf.data.key],
+                    repository_full_name=row[ghoprf.repository_full_name.key],
+                    author=authors[node_id])
+        return facts
+
+    @classmethod
+    @sentry_span
+    async def load_open_pull_request_facts_unfresh(cls,
+                                                   prs: Iterable[str],
+                                                   time_from: datetime,
+                                                   time_to: datetime,
+                                                   exclude_inactive: bool,
+                                                   authors: Mapping[str, str],
+                                                   account: int,
+                                                   pdb: databases.Database,
+                                                   ) -> Dict[str, PullRequestFacts]:
+        """
+        Fetch precomputed facts about the open PRs from the DataFrame.
+
+        We don't filter PRs by the last update here.
+
+        :param authors: Map from PR node IDs to their author logins.
+        :return: Map from PR node IDs to their facts.
+        """
+        postgres = pdb.url.dialect in ("postgres", "postgresql")
+        ghoprf = GitHubOpenPullRequestFacts
+        selected = [ghoprf.pr_node_id, ghoprf.repository_full_name, ghoprf.data]
+        default_version = ghoprf.__table__.columns[ghoprf.format_version.key].default.arg
+        filters = [
+            ghoprf.pr_node_id.in_(prs),
+            ghoprf.format_version == default_version,
+            ghoprf.acc_id == account,
+        ]
+        if exclude_inactive:
+            date_range = append_activity_days_filter(
+                time_from, time_to, selected, filters, ghoprf.activity_days, postgres)
+        rows = await pdb.fetch_all(select(selected).where(and_(*filters)))
+        if not rows:
+            return {}
+        facts = {}
+        for row in rows:
+            if exclude_inactive and not postgres:
+                activity_days = {datetime.strptime(d, "%Y-%m-%d").replace(tzinfo=timezone.utc)
+                                 for d in row[ghoprf.activity_days.key]}
+                if not activity_days.intersection(date_range):
+                    continue
+            node_id = row[ghoprf.pr_node_id.key].rstrip()
+            facts[node_id] = PullRequestFacts(
+                data=row[ghoprf.data.key],
+                repository_full_name=row[ghoprf.repository_full_name.key],
+                author=authors[node_id])
+        return facts
+
+    @classmethod
+    @sentry_span
+    async def load_open_pull_request_facts_all(cls,
+                                               repos: Collection[str],
+                                               pr_node_id_blacklist: Collection[str],
+                                               account: int,
+                                               pdb: databases.Database,
+                                               ) -> Dict[str, PullRequestFacts]:
+        """
+        Load the precomputed open PR facts through all the time.
+
+        We do not load the repository and the author!
+
+        :return: Map from PR node IDs to their facts.
+        """
+        ghoprf = GitHubOpenPullRequestFacts
+        selected = [
+            ghoprf.pr_node_id,
+            ghoprf.data,
+        ]
+        default_version = ghoprf.__table__.columns[ghoprf.format_version.key].default.arg
+        filters = [
+            ghoprf.pr_node_id.notin_(pr_node_id_blacklist),
+            ghoprf.repository_full_name.in_(repos),
+            ghoprf.format_version == default_version,
+            ghoprf.acc_id == account,
+        ]
+        query = select(selected).where(and_(*filters))
+        with sentry_sdk.start_span(op="load_open_pull_request_facts_all/fetch"):
+            rows = await pdb.fetch_all(query)
+        facts = {row[ghoprf.pr_node_id.key]: PullRequestFacts(row[ghoprf.data.key])
+                 for row in rows}
+        return facts
+
+
+@sentry_span
+async def store_open_pull_request_facts(
+        open_prs_and_facts: Iterable[Tuple[MinedPullRequest, PullRequestFacts]],
+        account: int,
+        pdb: databases.Database) -> None:
+    """
+    Persist the facts about open pull requests to the database.
+
+    Each passed PR must be open, we raise an assertion otherwise.
+    """
+    postgres = pdb.url.dialect in ("postgres", "postgresql")
+    if not postgres:
+        assert pdb.url.dialect == "sqlite"
+    values = []
+    for pr, facts in open_prs_and_facts:
+        assert not facts.closed
+        updated_at = pr.pr[PullRequest.updated_at.key]
+        if updated_at != updated_at:
+            continue
+        values.append(GitHubOpenPullRequestFacts(
+            acc_id=account,
+            pr_node_id=pr.pr[PullRequest.node_id.key],
+            repository_full_name=pr.pr[PullRequest.repository_full_name.key],
+            pr_created_at=pr.pr[PullRequest.created_at.key],
+            number=pr.pr[PullRequest.number.key],
+            pr_updated_at=updated_at,
+            activity_days=collect_activity_days(pr, facts, not postgres),
+            data=facts.data,
+        ).create_defaults().explode(with_primary_keys=True))
+    if postgres:
+        sql = postgres_insert(GitHubOpenPullRequestFacts)
+        sql = sql.on_conflict_do_update(
+            constraint=GitHubOpenPullRequestFacts.__table__.primary_key,
+            set_={
+                GitHubOpenPullRequestFacts.pr_updated_at.key: sql.excluded.pr_updated_at,
+                GitHubOpenPullRequestFacts.updated_at.key: sql.excluded.updated_at,
+                GitHubOpenPullRequestFacts.data.key: sql.excluded.data,
+            },
+        )
+    else:
+        sql = insert(GitHubOpenPullRequestFacts).prefix_with("OR REPLACE")
+    with sentry_sdk.start_span(op="store_open_pull_request_facts/execute"):
+        if pdb.url.dialect == "sqlite":
+            async with pdb.connection() as pdb_conn:
+                async with pdb_conn.transaction():
+                    await pdb_conn.execute_many(sql, values)
+        else:
+            # don't require a transaction in Postgres, executemany() is atomic in new asyncpg
+            await pdb.execute_many(sql, values)
